# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "The WIZARD Callable API",
    "version": "0.1.0",
    "summary": "Generated contract for callable runtime, planned DB-backed, accepted no-op, and DEV mock APIs.",
    "description": "This specification is generated from docs/api/openapi-source.json. DEV mock operations and accepted no-op placeholders are labeled with x-wizard-surface so they are not mistaken for production-ready DB-backed APIs."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/workflows": {
      "get": {
        "tags": [
          "Workflow Runtime"
        ],
        "operationId": "listWorkflows",
        "summary": "List workflow run projections once database-backed workflow reads are implemented.",
        "description": "List workflow run projections once database-backed workflow reads are implemented.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-wizard-surface": "db-backed-runtime-planned",
        "x-wizard-phase": "phase-0",
        "x-wizard-auth": "staff-session-required-planned",
        "x-wizard-write-boundary": "read-only"
      }
    },
...
    "/api/v1/approvals": {
      "get": {
        "tags": [
          "Workflow Runtime"
        ],
        "operationId": "listApprovals",
        "summary": "List approval tasks once database-backed approval reads are implemented.",
        "description": "List approval tasks once database-backed approval reads are implemented.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-wizard-surface": "db-backed-runtime-planned",
        "x-wizard-phase": "phase-0",
        "x-wizard-auth": "staff-session-required-planned",
        "x-wizard-write-boundary": "read-only"
      }
    },
...
    "/api/v1/annual-reset": {
      "post": {
        "tags": [
          "Workflow Runtime"
        ],
        "operationId": "createAnnualReset",
        "summary": "Accept an annual reset request without archiving rows in this checkout.",
        "description": "Accept an annual reset request without archiving rows in this checkout.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnualResetRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnualResetResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-wizard-surface": "accepted-no-op",
        "x-wizard-phase": "phase-0",
        "x-wizard-auth": "it-admin-required-planned",
        "x-wizard-write-boundary": "planned-db-write-boundary"
      }
    },
    "/api/v1/breakglass/login": {
      "post": {
        "tags": [
          "Breakglass"
        ],
        "operationId": "createBreakglassLogin",
        "summary": "Issue a local breakglass session after configured account, token, CIDR, and audit checks pass.",
        "description": "Issue a local breakglass session after configured account, token, CIDR, and audit checks pass.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BreakglassLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevSessionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [],
        "x-wizard-surface": "db-backed-runtime-conditional",
        "x-wizard-phase": "phase-0",
        "x-wizard-auth": "local-breakglass-token",
        "x-wizard-write-boundary": "session-and-audit-write"
      }
    },
...
    "/api/v1/dev/login": {
      "post": {
        "tags": [
          "DEV Mock APIs"
        ],
        "operationId": "createDevLogin",
        "summary": "Set the local DEV persona cookie and optional process-local tooling persona override.",
        "description": "Set the local DEV persona cookie and optional process-local tooling persona override.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevSessionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [],
        "x-wizard-surface": "dev-mock",
        "x-wizard-phase": "pre-phase-0",
        "x-wizard-auth": "development-only",
        "x-wizard-write-boundary": "cookie-and-dev-memory-write"
      }
    },
...
    "/api/v1/dev/my-profile": {
      "get": {
        "tags": [
          "DEV Mock APIs"
        ],
        "operationId": "getDevMyProfile",
        "summary": "Return mock self-service profile fields for the current DEV persona.",
        "description": "Return mock self-service profile fields for the current DEV persona.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevMyProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-wizard-surface": "dev-mock",
        "x-wizard-phase": "pre-phase-0",
        "x-wizard-auth": "dev-staff-session-required",
        "x-wizard-write-boundary": "read-only"
      },
      "put": {
        "tags": [
          "DEV Mock APIs"
        ],
        "operationId": "updateDevMyProfile",
        "summary": "Update allowed mock self-service profile fields for the current DEV persona.",
        "description": "Update allowed mock self-service profile fields for the current DEV persona.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevMyProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevMyProfileResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-wizard-surface": "dev-mock",
        "x-wizard-phase": "pre-phase-0",
        "x-wizard-auth": "dev-staff-session-required",
        "x-wizard-write-boundary": "dev-memory-write"
      }
    },
    "/api/v1/dev/feature-flags": {
      "get": {
        "tags": [
          "DEV Mock APIs"
        ],
        "operationId": "listDevFeatureFlags",
        "summary": "List DEV feature flags and target rows for IT Admin.",
        "description": "List DEV feature flags and target rows for IT Admin.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevFeatureFlagsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-wizard-surface": "dev-mock-or-local-db",
        "x-wizard-phase": "pre-phase-0",
        "x-wizard-auth": "it-admin-required",
        "x-wizard-write-boundary": "read-only"
      }
    },
...
    "/api/v1/dev/search": {
      "get": {
        "tags": [
          "DEV Mock APIs"
        ],
        "operationId": "searchDevGlobal",
        "summary": "Return route-authorized mock global search results.",
        "description": "Return route-authorized mock global search results.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevPageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-wizard-surface": "dev-mock",
        "x-wizard-phase": "pre-phase-0",
        "x-wizard-auth": "dev-staff-session-required",
        "x-wizard-write-boundary": "read-only"
      }
    },
    "/api/v1/dev/pages/onboarding": {
      "get": {
        "tags": [
          "DEV Mock APIs"
        ],
        "operationId": "getDevOnboardingPage",
        "summary": "Return the DEV onboarding page payload for authorized personas and site scope.",
        "description": "Return the DEV onboarding page payload for authorized personas and site scope.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevPageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-wizard-surface": "dev-mock",
        "x-wizard-phase": "pre-phase-0",
        "x-wizard-auth": "dev-route-permission-required",
        "x-wizard-write-boundary": "read-only"
      }
    },
    "/api/v1/dev/pages/offboarding": {
      "get": {
        "tags": [
          "DEV Mock APIs"
        ],
        "operationId": "getDevOffboardingPage",
        "summary": "Return the DEV offboarding page payload with persona-specific field visibility.",
        "description": "Return the DEV offboarding page payload with persona-specific field visibility.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevPageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-wizard-surface": "dev-mock",
        "x-wizard-phase": "pre-phase-0",
        "x-wizard-auth": "dev-route-permission-required",
        "x-wizard-write-boundary": "read-only"
      }
    },
    "/api/v1/dev/pages/reports/security-issues": {
      "get": {
        "tags": [
          "DEV Mock APIs"
        ],
        "operationId": "getDevSecurityIssuesReportPage",
        "summary": "Return the DEV IT Admin security-issues report payload.",
        "description": "Return the DEV IT Admin security-issues report payload.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevPageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-wizard-surface": "dev-mock",
        "x-wizard-phase": "pre-phase-0",
        "x-wizard-auth": "dev-it-admin-required",
        "x-wizard-write-boundary": "read-only"
      }
    },
    "/api/v1/dev/pages/departing-seniors": {
      "get": {
        "tags": [
          "DEV Mock APIs"
        ],
        "operationId": "getDevDepartingSeniorsPage",
        "summary": "Return the DEV departing-seniors page payload inside the retained senior-year window.",
        "description": "Return the DEV departing-seniors page payload inside the retained senior-year window.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevPageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-wizard-surface": "dev-mock",
        "x-wizard-phase": "pre-phase-0",
        "x-wizard-auth": "dev-it-or-device-wrangler-required",
        "x-wizard-write-boundary": "read-only"
      }
    },
    "/api/v1/dev/pages/data-quality": {
      "get": {
        "tags": [
          "DEV Mock APIs"
        ],
        "operationId": "getDevDataQualityPage",
        "summary": "Return the DEV IT Admin data-quality page payload.",
        "description": "Return the DEV IT Admin data-quality page payload.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevPageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-wizard-surface": "dev-mock",
        "x-wizard-phase": "pre-phase-0",
        "x-wizard-auth": "dev-it-admin-required",
        "x-wizard-write-boundary": "read-only"
      }
    },
    "/api/v1/dev/pages/meraki-last-seen": {
      "get": {
        "tags": [
          "DEV Mock APIs"
        ],
        "operationId": "getDevMerakiLastSeenPage",
        "summary": "Return the DEV Meraki last-seen dashboard payload with assignment-type and site-scope filtering.",
        "description": "Return the DEV Meraki last-seen dashboard payload with assignment-type and site-scope filtering.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevPageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-wizard-surface": "dev-mock",
        "x-wizard-phase": "pre-phase-0",
        "x-wizard-auth": "dev-it-site-admin-or-device-wrangler-required",
        "x-wizard-write-boundary": "read-only"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wizard_session",
        "description": "Local session cookie used by DEV and breakglass flows; production SAML wiring remains environment-managed."
      }
    },
    "schemas": {
...
      "AnnualResetRequest": {
        "type": "object",
        "additionalProperties": true,
        "properties": {}
      },
      "AnnualResetResponse": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "status": {
            "type": "string",
            "examples": [
              "accepted"
            ]
          },
          "workflow_type": {
            "type": "string",
            "examples": [
              "annual_reset_archive"
            ]
          }
        }
      },
...
      "BreakglassLoginRequest": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "account_id": {
            "type": "string"
          },
          "token": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "CollectionResponse": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "items": {
            "type": "array",
            "items": {}
          }
        }
      },
...
      "DevSessionResponse": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "authenticated": {
            "type": "boolean"
          },
          "authorized": {
            "type": "boolean"
          },
          "current_persona": {},
          "allowed_routes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request body, query string, or path parameter was invalid for this operation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  }
}